Added support for hiding YAML in PHP files to the YAML component.
* This is useful if you need to deploy an application to a cheap
  shared web hoster where you only have publically accessible
  directories and cannot use .htaccess to deny reading of yaml files
* __halt_compiler() and __COMPILER_HALT_OFFSET are used to extract
  the yaml data, as known from phar files.

I am not sure whether adding a dumpPHP function is really better than
adding a parameter for this to the existing dump() function.

I have added a copyright notice to the test files and an @author
tag to the dumpPHP function. I'm not sure what the correct process is
for symfony.